The module 'gmm' contains a function 'gaussian' to evaluate a multivariate normal distribution given a Dx1 feature vector, a Dx1 mean vector and a DxD covariance matrix.
The function'create_gmm' generates a GMM with M mixtures. Each mixture has a weight, a Dx1 feature vector and a DxD covariance matrix. The GMM is a list of length M, where each element is a tuple (weight, means, covariances).

The function 'eval_gmm' evaluates a given GMM for a Dx1 feature vector.

Signed-off-by: Eduardo Tenorio <embatbr@gmail.com>